requirements.txt
absl-py==2.2.2
astunparse==1.6.3
biopython==1.83
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.1.1
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5 
joblib==1.4.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pandas==2.0.3
pillow==10.4.0
protobuf==3.20.3
pyaml==25.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.3.2
scikit-optimize==0.10.2
scipy==1.10.1
six==1.17.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
